requirements.txt
telethon
flask
python-binance